// src/app/models/entities/Entidades.ts
export interface SetPersonasCommand {
  idPersonas?: number | undefined;
  identificacion?: string | undefined;
  nombres?: string | undefined;
  apellidos?: string | undefined;
  numeroCelular?: string | undefined;
  correoElectronico?: string | undefined;
  direccion?: string | undefined;
  sexo?: string | undefined;
  esActivo?: boolean;
}

export interface SetUsuarioCommand {
  idUsuario?: number | undefined;
  idPersonas?: number;
  idRol?: number | undefined;
  nombreUsuario?: string | undefined;
  clave?: string | undefined;
  esActivo?: boolean;
}